Edit double to float

+++ lab06/func.c
@@ -2,7 +2,7 @@
 # include <stdio.h>
 # include <math.h>
 
-double func (double x)
+float func (float x)
 {
   return(exp(-2*x*x-3)-x);  
 /*  return (-1 + 3*x + 5*x*x); */
